+++ lib/mokio/concerns.rb
@@ -50,6 +50,7 @@ require "mokio/concerns/models/base_contact"
 require "mokio/concerns/common/controller_object"
 require "mokio/concerns/common/controller_functions"
 require "mokio/concerns/common/controller_translations"
+require "mokio/concerns/common/services/sitemap_service"
 
 #
 # Controllers